dcp-815 separate class for metadata/provenance.py
